Use initContainer to check if the database is reachable
While developing the MariaDB integration, it happened to me several
times that MariaDB takes a while to become available, meanwhile the LS
container is waiting, but then get's shot down again by Kubernetes due
to the liveness probe.
This is especially bad when it happens during the database
initialization or while running migrations.
Thus, we just a separate init container, which will wait forever
without being killed.

Once the upstream LS image has a separate wait-for-db.sh command, we
can use that instead of specifying the shell command directly.