Use "main" branch version of chart-releaser-action
There was a bug that prevented the chart-releaser-action from working
with the current directory (".").
This bug was fixed in https://github.com/helm/chart-releaser/pull/83
but no new release has been tagged yet.

Thus, until a new release is out we can either just wait, or use the
branch "main" version in the meantime (though we shouldn't forget to
switch back to a tagged release once that is out).